Fixed the file loading situation.
I adjusted the file accessing, and added a public static parameter for easy access to the
target directory in case it changes.

Resources were moved to src/main/resources to better align with Maven.

The changed files now all use the public static BUILDDIR parameter,
which is set to the target directory based on our current structure. For
now I just put the parameter on the main class, since this isn't C or
C++.

There were some files that weren't actually being loaded correctly,
mostly the character and tile sprites not part of the world, so I made
those use ImageIO with Files. I also got the sounds to load correctly.
AudioManager now just loads the sounds from a File object instead of a
URL object since they are local.